Move classification engine namespace
