+++ lib/hackney/income/stored_tenancies_gateway.rb
@@ -9,7 +9,7 @@ module Hackney
 
         documents = DocumentModel.exclude_uploaded.by_payment_ref(criteria.payment_ref)
 
-        classification_usecase = Hackney::Income::TenancyClassification.new(
+        classification_usecase = Hackney::Income::TenancyClassification::Classifier.new(
           gateway_model_instance,
           criteria,
           documents
