+++ spec/lib/hackney/income/tenancy_classification/v1/classifier_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
-describe Hackney::Income::TenancyClassificationV1 do
+describe Hackney::Income::TenancyClassification::V1::Classifier do
   it_behaves_like 'TenancyClassification Contract'
 end
