+++ spec/lib/hackney/income/tenancy_classification/v2/classifier_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
-describe Hackney::Income::TenancyClassificationV2 do
+describe Hackney::Income::TenancyClassification::V2::Classifier do
   it_behaves_like 'TenancyClassification Contract'
 end
